+++ src/platform_impl/android/mod.rs
@@ -932,6 +932,9 @@ impl Window {
         ))
     }
 
+    #[inline]
+    pub fn show_system_menu(&self) {}
+
     pub fn set_cursor_hittest(&self, _hittest: bool) -> Result<(), error::ExternalError> {
         Err(error::ExternalError::NotSupported(
             error::NotSupportedError::new(),