+++ src/platform_impl/linux/mod.rs
@@ -440,6 +440,9 @@ impl Window {
     }
 
     #[inline]
+    pub fn show_system_menu(&self) {}
+
+    #[inline]
     pub fn set_cursor_hittest(&self, hittest: bool) -> Result<(), ExternalError> {
         x11_or_wayland!(match self; Window(w) => w.set_cursor_hittest(hittest))
     }