+++ src/platform_impl/macos/window.rs
@@ -892,6 +892,9 @@ impl WinitWindow {
     }
 
     #[inline]
+    pub fn show_system_menu(&self) {}
+
+    #[inline]
     pub fn set_cursor_hittest(&self, hittest: bool) -> Result<(), ExternalError> {
         self.setIgnoresMouseEvents(!hittest);
         Ok(())