+++ src/window.rs
@@ -1396,6 +1396,18 @@ impl Window {
             .maybe_wait_on_main(|w| w.drag_resize_window(direction))
     }
 
+    /// Show system menu on current mouse position.
+    ///
+    /// There's no guarantee that calling this method show the system menu.
+    /// user can cancel opening of it before it even open.
+    ///
+    /// ## Platform-specific
+    ///
+    /// calling this method **won't** do anything on any platform other then windows.
+    pub fn show_system_menu(&self) {
+        self.window.maybe_wait_on_main(|w| w.show_system_menu())
+    }
+
     /// Modifies whether the window catches cursor events.
     ///
     /// If `true`, the window will catch the cursor events. If `false`, events are passed through